+++ src/main/java/com/duboribu/ecommerce/front/order/dto/FoOrderItemView.java
@@ -4,8 +4,6 @@ import com.duboribu.ecommerce.front.enums.State;
 import com.querydsl.core.annotations.QueryProjection;
 import lombok.Getter;
 import lombok.ToString;
-
-import java.math.BigDecimal;
 @Getter
 @ToString
 public class FoOrderItemView {
@@ -24,13 +22,13 @@ public class FoOrderItemView {
     private String weight;
     @QueryProjection
     public FoOrderItemView(Long id, String title, String author, String publisher, String filePath,
-                      BigDecimal price, int stockCount, State state, String comment, String information, String weight) {
+                      int price, int stockCount, State state, String comment, String information, String weight) {
         this.id = id;
         this.title = title;
         this.author = author;
         this.publisher = publisher;
         this.filePath = filePath;
-        this.price = price.intValue();
+        this.price = price;
         this.stockCount = stockCount;
         this.state = state.name();
         this.comment = comment;
